requirements.txt
flask
langchain
langchain-google-genai
python-dotenv
